Move `Game` to its own file
